fix(senpi-task): harden plan-review target selection and split oversized test files
F2 review delta: explicit plan paths now only select session-recorded references (recorded path is interpolated, never the caller token); scope denial precedes one-shot classification; spawn policy composed in spawn-policy.ts (execute.ts back under 250 pure LOC); one-shot/result-renderer/reference-counting suites split into focused files; denial assertion uses registry equality instead of a phrase pin.
