+++ packages/senpi-task/src/tools/task/execute.ts
@@ -7,8 +7,7 @@ import { executeBatch } from "./execute-batch"
 import { buildStartSpec, singleSpawnParams } from "./execute-spec"
 import type { ForegroundWaitOptions } from "./foreground-wait"
 import { waitForForegroundTask } from "./foreground-wait"
-import { invocationGateDenial } from "./invocation-gate"
-import { planReviewContractOutcome } from "./plan-review-contract"
+import { evaluateSpawnPolicy } from "./spawn-policy"
 import type { TaskToolParamsStatic } from "./params"
 import { partialDetails, recordDetails, startedDetails, type SingleSpawnParams } from "./result-details"
 import { backgroundConversionText, backgroundStartText } from "./start-presentation"
@@ -58,21 +57,13 @@ async function runSpawn(
   if (selection.kind === "error") {
     return result(selection.error.message, { task_id: "", status: "invalid_arguments", mode: "spawn", reason: selection.error.message })
   }
-  let effectiveParams = params
-  if (selection.kind === "subagent_type") {
-    const sessionId = ctx.sessionManager.getSessionId()
-    const denial = invocationGateDenial(deps, selection.subagentType, sessionId)
-    if (denial !== undefined) {
-      return result(denial, { task_id: "", status: "denied", mode: "spawn", reason: denial })
-    }
-    const contract = planReviewContractOutcome(deps, selection.subagentType, params.prompt, sessionId)
-    if (contract?.kind === "deny") {
-      return result(contract.message, { task_id: "", status: "denied", mode: "spawn", reason: contract.message })
-    }
-    if (contract?.kind === "prompt") {
-      effectiveParams = { ...params, prompt: contract.prompt, load_skills: [] }
-    }
+  const policy = selection.kind === "subagent_type"
+    ? evaluateSpawnPolicy(deps, selection.subagentType, params.prompt, ctx.sessionManager.getSessionId())
+    : undefined
+  if (policy?.kind === "deny") {
+    return result(policy.message, { task_id: "", status: "denied", mode: "spawn", reason: policy.message })
   }
+  const effectiveParams = policy?.kind === "force" ? { ...params, prompt: policy.prompt, load_skills: [] } : params
   const target = selection.kind === "category" ? { category: selection.category } : { subagentType: selection.subagentType }
   const spec = buildStartSpec(effectiveParams, target, ctx.sessionManager.getSessionId(), deps, ctx.cwd)
   const started = await deps.manager.start(spec)
@@ -259,16 +250,12 @@ export function buildTaskExecute(deps: TaskToolDeps, options: ForegroundWaitOpti
         let itemParams = singleSpawnParams(item, params.run_in_background)
         const target = item.kind === "category" ? { category: item.category } : { subagentType: item.subagentType }
         if (item.kind === "subagent_type") {
-          const denial = invocationGateDenial(deps, item.subagentType, parentSessionId)
-          if (denial !== undefined) {
-            return { kind: "plan_unresolved", error: { code: "invalid_target", message: denial } }
+          const policy = evaluateSpawnPolicy(deps, item.subagentType, itemParams.prompt, parentSessionId)
+          if (policy.kind === "deny") {
+            return { kind: "plan_unresolved", error: { code: "invalid_target", message: policy.message } }
           }
-          const contract = planReviewContractOutcome(deps, item.subagentType, itemParams.prompt, parentSessionId)
-          if (contract?.kind === "deny") {
-            return { kind: "plan_unresolved", error: { code: "invalid_target", message: contract.message } }
-          }
-          if (contract?.kind === "prompt") {
-            itemParams = { ...itemParams, prompt: contract.prompt, load_skills: [] }
+          if (policy.kind === "force") {
+            itemParams = { ...itemParams, prompt: policy.prompt, load_skills: [] }
           }
         }
         const spec = buildStartSpec(itemParams, target, parentSessionId, deps, ctx.cwd)
